[DX] Move AddLocaleListener to EnrichBundle (only used fo UI)
