+++ spec/Pim/Bundle/EnrichBundle/EventListener/UserContextListenerSpec.php
@@ -4,7 +4,7 @@ namespace spec\Pim\Bundle\EnrichBundle\EventListener;
 
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Context\CatalogContext;
-use Pim\Bundle\TranslationBundle\EventListener\AddLocaleListener;
+use Pim\Bundle\EnrichBundle\EventListener\AddLocaleListener;
 use Pim\Bundle\UserBundle\Context\UserContext;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\HttpKernel\HttpKernel;
