+++ src/Pim/Bundle/TranslationBundle/DependencyInjection/PimTranslationExtension.php
@@ -21,7 +21,6 @@ class PimTranslationExtension extends Extension
     {
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('form_types.yml');
-        $loader->load('listeners.yml');
         $loader->load('twig.yml');
     }
 }
