--- src/Pim/Bundle/TranslationBundle/Resources/config/listeners.yml
@@ -1,9 +1,0 @@
-parameters:
-    pim_translation.listener.add_locale.class: Pim\Bundle\TranslationBundle\EventListener\AddLocaleListener
-
-services:
-    # inject current locale into translatable entities
-    pim_translation.listener.add_locale:
-        class: %pim_translation.listener.add_locale.class%
-        tags:
-            - { name: doctrine.event_subscriber }
